[DF] feat(#547): cleanup typescript code
