+++ fixture/src/twitter/Twitter.tsx
@@ -9,7 +9,6 @@ import {
 import { BlankAreaEventHandler, FlashList } from "@shopify/flash-list";
 import { FlashListPerformanceView } from "@shopify/react-native-performance-lists-profiler";
 
-import { ScrollView } from "../ScrollView";
 import { DebugContext } from "../Debug";
 
 import TweetCell from "./TweetCell";
@@ -90,7 +89,6 @@ const Twitter = ({
           console.log(info);
         }}
         disableAutoLayout={disableAutoLayout}
-        renderScrollComponent={ScrollView}
       />
     </FlashListPerformanceView>
   );
